Add tests and refactor processNotification tests to individual files for clarity
